requirements.txt
pb-tool==3.1.0
PyQt5-stubs==5.15.6.0
qgis-stubs==0.2.0.post1
